Create navigation bar
In this commit I have done the responsive navigation bar.
for desktop it will show only the text but for the mobile view it will show icons.
Using java script on hover it will show different icon for that I have used addEventListener function.
I have used custom made svgs for the icons.
I am not 100% satisfied with look may be I will do some changes in the icons but for the mvp part I am sticking with this look for now.